Extract node handler into dedicated crate
