+++ datacake-cluster/src/statistics.rs
@@ -6,10 +6,10 @@ pub type Counter = AtomicU64;
 
 #[derive(Debug, Clone, Default)]
 /// Live metrics around the cluster system.
-pub struct ClusterStatistics(Arc<ClusterStatisticsInner>);
+pub struct SystemStatistics(Arc<SystemStatisticsInner>);
 
-impl Deref for ClusterStatistics {
-    type Target = ClusterStatisticsInner;
+impl Deref for SystemStatistics {
+    type Target = SystemStatisticsInner;
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -17,11 +17,7 @@ impl Deref for ClusterStatistics {
 }
 
 #[derive(Debug, Default)]
-pub struct ClusterStatisticsInner {
-    /// The number of currently alive members the node is aware of.
-    pub(crate) num_live_members: Counter,
-    /// The number of members the node currently believes is dead.
-    pub(crate) num_dead_members: Counter,
+pub struct SystemStatisticsInner {
     /// The number of synchronisation tasks that are currently running concurrently.
     pub(crate) num_ongoing_sync_tasks: Counter,
     /// The number of synchronisation tasks that took longer than the selected timeout.
@@ -30,21 +26,9 @@ pub struct ClusterStatisticsInner {
     pub(crate) num_failed_sync_tasks: Counter,
     /// The number of times the node has observed a remote keyspace change.
     pub(crate) num_keyspace_changes: Counter,
-    /// The number of data centers/availability zones the cluster belongs to.
-    pub(crate) num_data_centers: Counter,
 }
 
-impl ClusterStatisticsInner {
-    /// The number of currently alive members the node is aware of.
-    pub fn num_live_members(&self) -> u64 {
-        self.num_live_members.load(Ordering::Relaxed)
-    }
-
-    /// The number of members the node currently believes is dead.
-    pub fn num_dead_members(&self) -> u64 {
-        self.num_dead_members.load(Ordering::Relaxed)
-    }
-
+impl SystemStatisticsInner {
     /// The number of synchronisation tasks that are currently running concurrently.
     pub fn num_ongoing_sync_tasks(&self) -> u64 {
         self.num_ongoing_sync_tasks.load(Ordering::Relaxed)
@@ -63,10 +47,5 @@ impl ClusterStatisticsInner {
     /// The number of times the node has observed a remote keyspace change.
     pub fn num_keyspace_changes(&self) -> u64 {
         self.num_keyspace_changes.load(Ordering::Relaxed)
-    }
-
-    /// The number of data centers/availability zones the cluster belongs to.
-    pub fn num_data_centers(&self) -> u64 {
-        self.num_data_centers.load(Ordering::Relaxed)
     }
 }
