+++ datacake-cluster/tests/basic_connect.rs
@@ -6,7 +6,7 @@ use datacake_cluster::{
     ClusterOptions,
     ConnectionConfig,
     DCAwareSelector,
-    DatacakeCluster,
+    EventuallyConsistentStore,
 };
 
 #[tokio::test]
@@ -16,7 +16,7 @@ async fn test_basic_connect() -> anyhow::Result<()> {
     let addr = "127.0.0.1:8000".parse::<SocketAddr>().unwrap();
     let connection_cfg = ConnectionConfig::new(addr, addr, Vec::<String>::new());
 
-    let cluster = DatacakeCluster::connect(
+    let cluster = EventuallyConsistentStore::connect(
         "node-1",
         connection_cfg,
         MemStore::default(),
