+++ datacake-node/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "datacake-node"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+thiserror = "1"
+flume = "0.10.14"
+tonic = "0.8.0"
+prost = "0.11.0"
+async-trait = "0.1.59"
+tracing = "0.1"
+tokio-stream = "0.1.9"
+parking_lot = "0.12.1"
+rand = "0.8.5"
+futures = "0.3"
+crc32fast = "1.3.2"
+
+datacake-crdt = { version = "0.3", path = "../datacake-crdt" }
+chitchat = { version = "0.5.1", package  = "datacake-chitchat-fork" }
+tokio = { version = "1", default-features = false, features = ["sync", "time"] }
+
+[dev-dependencies]
+tracing-subscriber = "0.3.16"
+tokio = { version = "1", features = ["full"] }
+
+[build-dependencies]
+tonic-build = "0.8.0"
+prost-build = "0.11.1"
