+++ datacake-node/proto/datacake.proto
@@ -1,0 +1,22 @@
+syntax = "proto3";
+
+package datacake_api;
+
+service ChitchatTransport {
+    rpc send_msg(ChitchatRpcMessage) returns (Empty);
+}
+
+message ChitchatRpcMessage {
+    bytes message = 1;
+    bytes source = 2;
+    Timestamp timestamp = 3;
+}
+
+message Empty {};
+
+/// A HLCTimestamp.
+message Timestamp {
+  uint64 millis = 1;
+  uint32 counter = 2;
+  uint32 node_id = 3;
+}
