+++ datacake-node/src/rpc/network.rs
@@ -6,6 +6,7 @@ use std::time::Duration;
 
 use parking_lot::RwLock;
 use tonic::transport::{Channel, Endpoint, Error};
+use tracing::trace;
 
 pub const TIMEOUT_LIMIT: Duration = Duration::from_secs(2);
 pub const CONNECT_TIMEOUT_LIMIT: Duration = Duration::from_secs(2);
